+++ v3/crates/metadata-resolve/src/stages/model_permissions/model_permission.rs
@@ -10,7 +10,7 @@ use crate::helpers::typecheck::typecheck_value_expression;
 use crate::helpers::types::mk_name;
 use crate::stages::{
     boolean_expressions, data_connector_scalar_types, data_connectors, models, models_graphql,
-    object_boolean_expressions, object_types, relationships, scalar_types, type_permissions,
+    object_boolean_expressions, object_relationships, object_types, scalar_types, type_permissions,
 };
 use crate::types::error::{Error, TypePredicateError};
 use crate::types::permission::ValueExpression;
@@ -43,7 +43,10 @@ fn resolve_model_predicate_with_model(
         data_connector_scalar_types::ScalarTypeWithRepresentationInfoMap,
     >,
     fields: &IndexMap<FieldName, object_types::FieldDefinition>,
-    object_types: &BTreeMap<Qualified<CustomTypeName>, relationships::ObjectTypeWithRelationships>,
+    object_types: &BTreeMap<
+        Qualified<CustomTypeName>,
+        object_relationships::ObjectTypeWithRelationships,
+    >,
     scalar_types: &BTreeMap<Qualified<CustomTypeName>, scalar_types::ScalarTypeRepresentation>,
     boolean_expression_types: &boolean_expressions::BooleanExpressionTypes,
     models: &IndexMap<Qualified<ModelName>, models_graphql::ModelWithGraphql>,
@@ -142,7 +145,10 @@ pub fn resolve_model_select_permissions(
         Qualified<DataConnectorName>,
         data_connector_scalar_types::ScalarTypeWithRepresentationInfoMap,
     >,
-    object_types: &BTreeMap<Qualified<CustomTypeName>, relationships::ObjectTypeWithRelationships>,
+    object_types: &BTreeMap<
+        Qualified<CustomTypeName>,
+        object_relationships::ObjectTypeWithRelationships,
+    >,
     scalar_types: &BTreeMap<Qualified<CustomTypeName>, scalar_types::ScalarTypeRepresentation>,
     models: &IndexMap<Qualified<ModelName>, models_graphql::ModelWithGraphql>,
     object_boolean_expression_types: &BTreeMap<
@@ -265,13 +271,16 @@ pub fn resolve_model_select_permissions(
 pub(crate) fn resolve_model_predicate_with_type(
     model_predicate: &open_dds::permissions::ModelPredicate,
     type_name: &Qualified<CustomTypeName>,
-    object_type_representation: &relationships::ObjectTypeWithRelationships,
+    object_type_representation: &object_relationships::ObjectTypeWithRelationships,
     boolean_expression_graphql: Option<&boolean_expressions::BooleanExpressionGraphqlConfig>,
     data_connector_field_mappings: &BTreeMap<FieldName, object_types::FieldMapping>,
     data_connector_link: &data_connectors::DataConnectorLink,
     subgraph: &SubgraphName,
     scalars: &data_connector_scalar_types::ScalarTypeWithRepresentationInfoMap,
-    object_types: &BTreeMap<Qualified<CustomTypeName>, relationships::ObjectTypeWithRelationships>,
+    object_types: &BTreeMap<
+        Qualified<CustomTypeName>,
+        object_relationships::ObjectTypeWithRelationships,
+    >,
     scalar_types: &BTreeMap<Qualified<CustomTypeName>, scalar_types::ScalarTypeRepresentation>,
     boolean_expression_types: &boolean_expressions::BooleanExpressionTypes,
     models: &IndexMap<Qualified<ModelName>, models_graphql::ModelWithGraphql>,
@@ -437,21 +446,21 @@ pub(crate) fn resolve_model_predicate_with_type(
                     })?;
 
                 match &relationship.target {
-                    relationships::RelationshipTarget::Command { .. } => {
+                    object_relationships::RelationshipTarget::Command { .. } => {
                         Err(Error::UnsupportedFeature {
                             message: "Predicate cannot be built using command relationships"
                                 .to_string(),
                         })
                     }
-                    relationships::RelationshipTarget::ModelAggregate { .. } => {
+                    object_relationships::RelationshipTarget::ModelAggregate { .. } => {
                         Err(Error::UnsupportedFeature {
                             message:
                                 "Predicate cannot be built using model aggregate relationships"
                                     .to_string(),
                         })
                     }
-                    relationships::RelationshipTarget::Model(
-                        relationships::ModelRelationshipTarget {
+                    object_relationships::RelationshipTarget::Model(
+                        object_relationships::ModelRelationshipTarget {
                             model_name,
                             relationship_type,
                             target_typename,
@@ -844,7 +853,7 @@ fn resolve_binary_operator_for_type<'a>(
 fn remove_object_relationships(
     object_types_with_relationships: &BTreeMap<
         Qualified<CustomTypeName>,
-        relationships::ObjectTypeWithRelationships,
+        object_relationships::ObjectTypeWithRelationships,
     >,
 ) -> type_permissions::ObjectTypesWithPermissions {
     type_permissions::ObjectTypesWithPermissions(
@@ -853,7 +862,7 @@ fn remove_object_relationships(
             .map(
                 |(
                     object_name,
-                    relationships::ObjectTypeWithRelationships {
+                    object_relationships::ObjectTypeWithRelationships {
                         object_type,
                         type_mappings,
                         type_input_permissions,