+++ v3/crates/metadata-resolve/src/stages/relationships/error.rs
@@ -1,0 +1,19 @@
+use crate::types::subgraph::Qualified;
+use open_dds::relationships::RelationshipName;
+use open_dds::types::CustomTypeName;
+
+#[derive(Debug, thiserror::Error)]
+pub enum RelationshipError {
+    #[error("Relationship {relationship_name} could not be found for type {object_type_name}")]
+    RelationshipNotFound {
+        object_type_name: Qualified<CustomTypeName>,
+        relationship_name: RelationshipName,
+    },
+    #[error(
+        "Multiple relationships named {relationship_name} defined for type {object_type_name}"
+    )]
+    DuplicateRelationshipForType {
+        object_type_name: Qualified<CustomTypeName>,
+        relationship_name: RelationshipName,
+    },
+}