+++ v3/crates/metadata-resolve/src/stages/types.rs
@@ -15,14 +15,14 @@ use crate::types::subgraph::Qualified;
 
 use crate::stages::{
     aggregates, boolean_expressions, command_permissions, graphql_config, model_permissions,
-    object_boolean_expressions, order_by_expressions, relationships, scalar_types,
+    object_boolean_expressions, object_relationships, order_by_expressions, scalar_types,
 };
 
 /// Resolved and validated metadata for a project. Used internally in the v3 server.
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq)]
 pub struct Metadata {
     pub object_types:
-        BTreeMap<Qualified<CustomTypeName>, relationships::ObjectTypeWithRelationships>,
+        BTreeMap<Qualified<CustomTypeName>, object_relationships::ObjectTypeWithRelationships>,
     pub scalar_types: BTreeMap<Qualified<CustomTypeName>, scalar_types::ScalarTypeRepresentation>,
     pub models: IndexMap<Qualified<ModelName>, model_permissions::ModelWithPermissions>,
     pub commands: IndexMap<Qualified<CommandName>, command_permissions::CommandWithPermissions>,